LF: refactor java codegen integration tests
This advances the state of #7155.

CHANGELOG_BEGIN
CHANGELOG_END
