+++ language-support/java/codegen/BUILD.bazel
@@ -29,6 +29,7 @@ load(
 )
 load("//bazel_tools:pom_file.bzl", "pom_file")
 load("@os_info//:os_info.bzl", "is_windows")
+load("@build_environment//:configuration.bzl", "ghc_version", "sdk_version")
 
 da_scala_binary(
     name = "codegen",
@@ -163,55 +164,68 @@ test_daml_lf_target_versions = [
 ########################################################
 
 it_daml_lf_target_versions = [
-    "1.0",
-    "1.1",
-    "1.5",
-    "latest",
+    "1.6",
+    "1.7",
+    "1.8",
 ]
 
-# This file was created using a patched version of 0.12.12.
-# See https://github.com/digital-asset/daml/pull/2443 for more details
-alias(
-    name = "integration-tests-model-1.0.dar",
-    actual = ":src/it/dar/integration-tests-model-1.0.dar",
-)
-
-# This file was created using
-# DAML_SDK_VERSION=0.12.20 daml damlc package src/it/daml/Lib.daml integration-tests-model-1.1 --target=1.1 --output src/it/dar/integration-tests-model-1.1.dar
-alias(
-    name = "integration-tests-model-1.1.dar",
-    actual = ":src/it/dar/integration-tests-model-1.1.dar",
-)
-
-# This file was created using
-# DAML_SDK_VERSION=0.13.10 daml damlc package src/it/daml/Lib.daml integration-tests-model-1.5 --target=1.5 --output src/it/dar/integration-tests-model-1.5.dar
-alias(
-    name = "integration-tests-model-1.5.dar",
-    actual = ":src/it/dar/integration-tests-model-1.5.dar",
-)
-
-daml_compile(
-    name = "integration-tests-model-latest",
-    srcs = glob(["src/it/daml/**/*.daml"]),
-    target = lf_latest_version,
-)
-
 [
-    #
-    # Compile the DAR files with and without a package to allow
-    # DAML-LF version agnostic tests to be run using generated
-    # Java classes with the same qualified name, while the DAML-LF
-    # specific tests run using Java classes generated with a
-    # package name qualified by the DAML-LF version.
-    #
     [
+        genrule(
+            name = "integration-tests-model-%s" % mangle_for_java(target),
+            srcs = ["src/it/daml"],
+            outs = ["integration-tests-model-%s.dar" % mangle_for_java(target)],
+            cmd = """
+             set -eou pipefail
+             TMP_DIR=$$(mktemp -d)
+             cp -Lr $(location src/it/daml) $$TMP_DIR
+             cat << EOF > $$TMP_DIR/daml.yaml
+sdk-version: {sdk}
+name: daml-trigger
+source: daml
+version: {ghc}
+dependencies:
+  - daml-stdlib
+  - daml-prim
+build-options: {build_options}
+EOF
+             $(location //compiler/damlc) build --project-root $$TMP_DIR \
+                 --cpp $$PWD/$(location @stackage-exe//hpp) \
+                 -o $$PWD/$@
+             rm -rf $$TMP_DIR
+           """.format(
+                build_options = str([
+                    "--target",
+                    target,
+                ]),
+                sdk = sdk_version,
+                ghc = ghc_version,
+            ),
+            tools = [
+                "//compiler/damlc",
+                "@stackage-exe//hpp",
+            ],
+            visibility = ["//visibility:public"],
+        ),
+        #        daml_compile(
+        #            name = "integration-tests-model-%s" % mangle_for_java(target),
+        #            srcs = glob(["src/it/daml/**/*.daml"]),
+        #            target = target,
+        #        ),
+        #
+        # Compile the DAR files with and without a package to allow
+        # DAML-LF version agnostic tests to be run using generated
+        # Java classes with the same qualified name, while the DAML-LF
+        # specific tests run using Java classes generated with a
+        # package name qualified by the DAML-LF version.
+        #
         dar_to_java(
             name = "integration-tests-model-noprefix-%s" % target,
-            src = ":integration-tests-model-%s.dar" % target,
+            src = ":integration-tests-model-%s.dar" % mangle_for_java(target),
         ),
         dar_to_java(
             name = "integration-tests-model-%s" % target,
-            src = ":integration-tests-model-%s.dar" % target,
+            src = ":integration-tests-model-%s.dar" % mangle_for_java(target),
             package_prefix = "lf_%s" % mangle_for_java(target),
         ),
         java_test(
