--- language-support/java/codegen/src/it/java-1.0/com/daml/lf_1_0/AllTests.java
@@ -1,17 +1,0 @@
-// Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.lf_1_0;
-
-import com.daml.AllGenericTests;
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({
-        OptionalTest.class,
-        ListTest.class,
-        AllGenericTests.class
-})
-public class AllTests {
-}
