--- language-support/java/codegen/src/it/java-1.5/com/daml/lf_1_5/AllTests.java
@@ -1,19 +1,0 @@
-// Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.lf_1_5;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import com.daml.AllGenericTests;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({
-        AllGenericTests.class,
-        ListTest.class,
-        OptionalTest.class,
-        MapTest.class,
-        ContractKeysTest.class,
-})
-public class AllTests {
-}
