--- language-support/java/codegen/src/test/java/com/digitalasset/testing/AllTestsFor1_3.java
@@ -1,13 +1,0 @@
-// Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.testing;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({
-        DecimalTestForAll.class
-})
-public class AllTestsFor1_3 { }
